H3DHapticsDevice.h - Doxygen update.
Switch.cpp - Fixed bug 0000123 cause by a statement similar to VAR - 1 where VAR is an unsigned integer.

git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@1136 610dfd11-2700-0410-9b68-9ae28eac5983

+++ src/Switch.cpp
@@ -66,7 +66,7 @@ Switch::Switch( Inst< AddChildren    > _addChildren,
 
 void Switch::render() {
   int choice = whichChoice->getValue();
-  if( choice < 0 || (size_t)choice > children->size() - 1 ) return;
+  if( choice < 0 || choice > (int)children->size() - 1 ) return;
   X3DChildNode *child_node = children->getValueByIndex( choice );
   H3DDisplayListObject *dlo = 
     dynamic_cast< H3DDisplayListObject * >(child_node);
@@ -86,7 +86,7 @@ void Switch::SFBound::update() {
 
 void Switch::traverseSG( TraverseInfo &ti ) {
   int choice = whichChoice->getValue();
-  if( choice < 0 || (size_t)choice > children->size() - 1 ) return;
+  if( choice < 0 || choice > (int)children->size() - 1 ) return;
   X3DChildNode *child_node = children->getValueByIndex( choice );
   if( child_node ) child_node->traverseSG( ti );
 }
